1. NoteListScreen UI 화면 추가 2. 파일 구조 수정
