+++ app/src/main/AndroidManifest.xml
@@ -12,17 +12,6 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.JoopJoop" >
         <activity
-            android:name=".note.NoteListScreen"
-            android:exported="true"
-            android:label="@string/title_activity_write_note_screen"
-            android:theme="@style/Theme.JoopJoop" >
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
